update register to user model

+++ backend/users/views.py
@@ -10,10 +10,16 @@ class UserView(APIView):
         return Response([user.to_json() for user in users], status=status.HTTP_200_OK)
 
     def post(self, request):
-        d = request.data
-        user = User.from_json(d)
-        user.insert()
-        return Response(user.to_json(), status=status.HTTP_200_OK)
+        try:
+            d = request.data
+            user = User.from_json(d)
+            if not user.is_exists():
+                user.insert()
+                return Response(user.to_json(), status=status.HTTP_200_OK)
+            else:
+                return Response({"error": "user already exists"}, status=status.HTTP_400_BAD_REQUEST)
+        except Exception as e:
+            return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, id=None):
         if id is None:
